requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
GeoAlchemy2==0.14.2
flasgger==0.9.7.1

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Password hashing
bcrypt==4.1.2
